+++ CHANGELOG.md
@@ -10,6 +10,8 @@ about writing changes to this log.
 
 ## [Unreleased]
 
+- [PR-394](https://github.com/itk-dev/naevnssekretariatet/pull/394)
+  Added logout functionality.
 - [PR-393](https://github.com/itk-dev/naevnssekretariatet/pull/393)
   Fixed `BoardRole` crud search property.
 - [PR-392](https://github.com/itk-dev/naevnssekretariatet/pull/392)